requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
sentence-transformers
PyPDF2
requests
numpy
python-dotenv
python-multipart
loguru
httpx
pytest
pytest-asyncio